Solve Day 8 Part 2.
Turns out the way I was doing it was ridiculously inefficient, and it
wasn't about getting faster lookups. I tried with `array` and I got
the lookup times down to about two thirds, but that wasn't anywhere near
fast enough.

The real solution is to find out how many cycles it takes each path
individually to hit a Z node, and then find the lowest common multiple
of all of those numbers. Like working out when all the parts of a
polyrhythm will land on the same beat. Actually following all the paths
until they are all on a Z at the same time is a huge waste of time.