Optimise Day 12 some more.
This solution is way faster than my previous attempts, oddly enough by
being more dumb. Instead of trying to do clever tricks with
'combinations' functions, I just iterate through each ? position and
then recurse with a dot and with a hash, and add up all the outcomes
that match the pattern.

Because it stops trying a branch as soon as it fails the pattern, there
are many combinations that don't end up getting tried.