Further optimise Day 12 Part 2.
This attempt only calculates the number of solutions for each unique
grouping of contiguous copies of the row joined by a '#', and from there
it's just a maths formula to get the answer.

This solves the test case in < 0.2s which is a huge improvement, but it
is still unable to solve the full input in a reasonable time.