Work in progress for Day 17.
I got pretty close with a Dijkstra pathing attempt. The problem is the
restriction on going straight for more than 3 moves. With Dijkstra, we
only care about the shortest distance to each node, not about what
direction we are travelling to get there.

I got an answer of 100 with the first attempt, which is too low, should
be 102. Then I tried making the direction of travel part of the node
key, and that got me 105, so still not quite right.