Prep work for Day 20 Part 2.
I added methods to reset modules, and to detect when they are at their
initial state.

Still not really sure how I'm going to reason out the number of cycles
needed to produce a particular output. I suspect it's going to be a case
of combinatorial logic but it isn't obvious how to go about it.