"Solve" Day 20 Part 2.
This puzzle was completely different than all the others. Instead of
programming up a general solution for any input, the only sensible way
to do this Part 2 was to study the configuration of the puzzle input to
understand how it works, and then run some empirical tests to figure out
the cycle lengths. Gross.