Add a logically sound, but unperformant, solution for Day 24 Part 2.
This solution works very well for the test case, but has not been able
to solve the real input in reasonable time.

I am thinking about two possible ways to improve this. One is to test
the shortest distance between my candidate rock trajectory and the third
hailstone. If that distance is getting bigger as I advance t2, then
advancing it more will not help and I can abandon that t1 value
immediately.  The problem is, the third hailstone might be converging
while others are diverging, so I will have to test all hailstones to
figure out whether any of them are diverging.

The other option is to use a plane. If I take a starting t value, and
draw a line from the hailstone with the lowest Z to the one with the
highest Z at the starting time, and from the same lowest point to the
highest hailstone at an arbitrary later t, if any hailstone does not
pass through that plane, then the starting t value cannot ever provide a
solution.