Optimise Day 24 Part 2.
This version uses a 3D plane between the lowest stone and the highest
stone at a given time value, and rejects the time position if not all of
the hailstones have an intersection with that plane.

This works pretty fast, but still not fast enough for Part 2. I have let
it run through the first 300k or so time positions with no hits.