Move all 2023 solutions into their own directory.
The idea is to use a generic 'runner' script at the top level to load
the solution code from modules, instead of each solution being its own
script.

In this commit I've implemented the changes for 2023/12 but I still need
to rework all the other 2023 solutions to use the generic runner.