Optimise in 2015 Day 11.
I tried a few different things, but in the end none of it was massively
faster than the naive implementation for the test case 'ghijklmn'.

I settled on the bytearray version because it's tight and memory
efficient, even though it's not very much faster overall. And to avoid
bogging down the test runs, I replaced the test case with the much
faster option 'abcdefgh'.