Failed attempt to solve 2015 Day 19 Part 2.
Here I'm using an A* pathfinding approach to try to solve the problem. I
have optimised it to the point where it solves the test case in less
than a millisecond (!) but it still will not solve the actual case in a
reasonable time.

I have since learned that the problem is meant to be a context-free
grammar type of problem, so next I will try solving it with a CYK
algorithm.