Fail again to solve 2015 Day 19 Part 2.
Following a lead about context-free grammars from this post:

https://www.reddit.com/r/adventofcode/comments/16calg6/was_the_real_solution_to_2015_day_19_ever_revealed/

I went down the rabbit hole of writing a CYK algorithm, which required
me to also write a transforming function to put the production rules
from the puzzle into Chomsky Normal Form.

Boy did that turn out to be a huge waste of effort, because CYK
apparently only works on target strings that consist entirely of
terminal symbols. Which the puzzle molecule very definitely is not.

This code works very well on the example grammar from the Wikipedia page
on CYK but it doesn't work for the puzzle input, or even for the HOHOHO
test case.

I'm committing this in case I ever need to refer back to my solution,
for a situation that actually *does* call for CYK or Chomsky Normal
Form.