Fail yet again to solve 2015 Day 19 Part 2.
This time I tried the Earley Parser and got it working for a sample, but
it doesn't work with the puzzle.

As with CYK, I think this is because the puzzle molecule is a mix of
terminal and non-terminal symbols.

And likewise, I am committing this code in case I need an Earley Parser
at some later time. Compared to CYK, it's nice that I don't have to
convert the grammar into CNF, but the algorithm itself is a lot more
difficult to reason about.