Sort-of solve 2015 Day 19?
I mean this has got to be the saddest solve of the bunch. I got
absolutely sick of trying to understand how to extract parsing
information out of an Earley recogniser, gave up and decided to try just
replacing the longest possible match at each step to get from the target
back to the start string. And that worked ... some of the time. It
really depends on how the replacement rules happen to be sorted within
the set, which changes from one execution of Python to the next.

I literally just spammed the button until one pass succeeded and put the
answer into the Advent page.

This was gross and I hated it.