Solve 2015 Day 22 Part 2.
The problem with my logic was the selection of spells for the next
round. I was excluding all spells currently in effect, forgetting that
some of those effects will end at the start of the next round and those
spells can be re-cast. Excluding only those spells with effects that
have more than 1 round left on their timers was the ticket.