Solve 2016 Day 11, both parts.
However, Part 2 is painfully slow. It takes around 221s to run which is
not the kind of performance I like to see.

The test case still runs in a reasonable amount of time, but adding
those final two objects to the ground floor in the actual Part 2 really
blows it out.