Optimise 2016 Day 11.
I have switched to a BFS for the pathfinding. Not sure whether that was
actually useful, but I also changed the way I structure the items to
allow for a kind of cheat in is_safe().

That brings the total runtime down to 27s, nearly a 10x improvement.  I
think it could still be better, but it's definitely in the right
ballpark now.