Solve 2017 Day 15, both parts (but it's slow you guys).
It's so slow.

This solution takes about 21s to do Part 1 and slightly less to do Part
2. Yuck.

I'm sure there is a very smart big brain way to solve this with maths
but I don't even know where to start with that. I did not spot any
particular patterns in the timing of the matches.

I could arrange to decide both Part 1 and 2 results on the same
generation loop, and that will cut the time by some amount, but I still
feel like I'm missing out on some big gains somewhere.