Use Numba to improve performance in y2017d15.
So it turns out, there is no real big brain way to make this code
faster, it just happens to be the maximally bad use-case for an
interpreted language compared to a compiled one.

The only way to make it substantially faster is to use some kind of
compiling. I saw PyPy and Cython and Numba recommended. PyPy seemed like
a fun idea but it was a pain to get going with it, and my distro only
offers a package for 3.8. Cython seemed like it was going to be annoying
to implement, especially for automated test runs.

Numba was the answer. It's just a pip package install and then I get
speed improvements of around SIXTY TIMES on this puzzle.

Yes that's right, sixty times. From around 37s to around 0.6s.
Ridiculous.