Implement the smart solution to 2017 Day 17 Part 2.
Man I really feel dumb. I spent so much time trying to get this to run
efficiently, finally gave up and had a look on Reddit, only to see the
hint that I don't actually need the buffer.

Wait, what? No buffer? Oh right. The only time I need to care about a
value getting added to the buffer is when it happens to be right after
zero, which is something I can work out from just the size of the buffer
and the current position.