Preparation for 2017 Day 21.
I've got all the code to parse the transform rule set, all the code to
generate an efficient lookup for all possible subgrids of size 2 and 3,
and all the code to do flip and rotate operations on grids.

Now all I need to do is slicing and recombining grids according to the
puzzle rules and I should be done with Part 1.