Solve 2018 Day 15, Part 2.
I ended up in the awkward position where I had Part 1 correct, and all
of my tests were passing for both parts, but I was still getting the
wrong answer for Part 2.

It turned out that my effort to optimise select_move() by prioritising
targets by their manhattan distance was the problem. It meant that when
two routes had the exact same cost, I was breaking ties in favour of
manhattan distance, not by reading order. And apparently that
distinction never came up in any of the test runs or any of the Part 1
stuff.