Solve 2018 Day 23, Part 2.
OK so I did need a hint for this one. The strategy of slicing up a 3D
rectangular volume like a binary search was probably never going to
occur to me, and I'd never heard of a "branch and bound" optimisation
before.

I am pleased with the performance of my solution though, it really is
quick, ~350ms for Part 2. That is not bad for narrowing down a total
space that has probably quintillions of points in it.