Solve 2019 Day 12, Part 2.
I needed a little nudge in the right direction for this one. I was kinda
on the right track but had an off-by-one error that was throwing me off
the scent.

The trick was to treat X, Y and Z each as independent variables, find
the cycles in each axis and get the lowest common multiple of those
cycles.