Solve 2019 Day 13, Part 2.
This was an interesting one. I could have set it up so that I actually
play the game myself, but I didn't want to pull in external libraries to
read the keyboard input, and doing it turn-based didn't seem like fun.

So instead, I wrote some code to play the game automatically. It worked
very smoothly.