Progress on 2019 Day 16, Part 2.
After some optimising, this code can now produce the correct answer for
one of the Part 2 test cases in 60s.  That's still too slow but it is
definitely a step in the right direction.

The tricks I've used so far are:

- Modify the signal list in-place, rather than making copies
- Work out the positive and negative slices from the pattern, and sum
  the digits in those slices, rather than multiplying by 1, 0 or -1.
  This means we are not doing any operation on the regions covered by
  zeroes in the pattern.
- Don't bother calculating the elements prior to the message offset.
  Because of the way the pattern progresses, the part of the signal
  prior to the offset cannot impact on the results after the offset.

I suspect there are more tricks to leverage here. Maybe there are cycles
that we can detect, or ways we can ignore later digits in the pattern.