Set up for 2019 Day 18.
I thought I had a working solution here, it solves the original test
case quite swiftly. But against the 16 key example, it is a shambles.

I think I need to find some way to cut down on the number of branches.

The inputs are all made of up of very narrow tunnels, so it might be
efficient to compute a graph of all the intersections/doors/keys and do
pathfinding on those, instead of pathfinding through each square every
time.