Solve 2019 Day 18, Part 1.
It's wild how long I spent mucking about with this one, now that I look
at the solution it seems so obvious.

The main issue was that my pathfinding problem was too large, and I was
making it hard on myself by not pruning out moves with equivalent
outcomes. The answer was to use Dijkstra with the nodes being the
unordered set of keys collected, plus the position. That effectively
removes paths which end up back at the same location with the same set
of keys, for the same cost, just collected in a slightly different
order.