Solve 2019 Day 22, Part 1.
I have no idea how I'm going to approach Part 2. I thought about tracing
the position of the target card back through each of the instructions,
but I don't know how to do that for the deal with increments, and I
don't see how it could help with the insane number of repeat shuffles
required.