Set up for 2020 Day 18 part 2.
I sunk a bunch of time into trying to write a generic context-free
grammar module, but in the end it wasn't that useful for the problem I
was trying to solve.

I did end up using a predictive parser based on a grammar, but it uses a
hardcoded parsing program rather than trying to run the grammar in a
general way.

I should be able to adapt this logic to work with the new precedence
rules in part 2.