Solve 2020 Day 19, Part 2.
The problem with the previous solution was balancing the number of 42s
and 31s in a rule 11 match.

Regex is a tool not well suited to this task, but I really didn't want
to cook up some kind of grammar-based parser for this, so I used a nasty
compromise solution of adding branches with each possible number of
matching 42 and 31 groups, adding more branches until the results didn't
change. By trial, I only needed 4 branches to reach this point.

Kinda gross but it did the job.