A spot of reorganising of the repo.
Rename the example input directories under each year to 'examples'
rather than 'tests', because 'tests' is also used with the unit test
stuff.

Rename newday.sh to setup.sh, because it's also useful to setup the
environment variables for a given puzzle even when not creating it for
the first time.