Solve 2020 Day 23, Part 2.
I needed a hint for this one. I was thinking it was something like
finding a cycle, or applying some modular arithmetic nonsense. But in
fact it was just a plain old optimisation problem.

The thing that slows us down in Part 2 is just traversing the list to
find the destination node in each round. We can shortcut this whole
thing but just having a static dict that maps node values to their
nodes, and then we jump directly to any node if we know its value.

Credit to https://www.reddit.com/r/adventofcode/comments/kiydf6/comment/ggtirpk/