Attempt 2021 Day 14 Part 2.
I changed to a depth-first iteration through the space, which is nice
and quick for 10 levels deep, and in theory will actually be able to
complete the 40 level problem in some kind of finite time. But it is
still visiting every node and with trillions of nodes to process, it's
not reasonable to visit them all.

I'm thinking we could build up a lookup of node and number of levels
remaining to a counter. That will be my next try.