Optimise in 2021 Day 23.
I've changed the internal structure for the board state to use columns.
This makes it a lot quicker to iterate through the valid pods to move,
because only the topmost pod in any given column can possibly move.

I've also enhanced the pathfinding heuristic to consider the minimum
number of moves for each of the pods that isn't home yet.

This gets the runtime for Part 1 test down to 600ms and Part 1 actual to
3.5s, which is a very considerable improvement.  Part 2 still doesn't
work.