Add feature to automatically create a puzzle program.
In interactive mode, when selecting a year/day (or starting up
interactive mode) advent.py will check to see whether a puzzle program
file already exists for that day.  If it doesn't, it will ask whether
you want to create it, and if you answer yes, will create both the
puzzle program file and the unit test case for that day.

With this feature, the interactive mode can now do everything that I
previously did with the setup.sh shell script and environment variables.