Resolve numba deprecation warnings in 2019 day 16 code.
Use of native Python lists and sets in numba-wrapped code is deprecated,
I resolved it by converting the Python list to a numpy array before
passing it into the numba code.

This fixed the deprecation warning and also delivered a nice little
performance boost as a side effect (~800ms -> ~500ms for both parts in
test mode).